scheduler/requirements.txt
openpyxl>=3.1
pyyaml>=6.0
pytest>=8.0
anthropic>=0.40
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.0
